+++ dosrtl/alloc.c
@@ -5,7 +5,7 @@ void *RtlAllocateMemory(
     size_t size)
 {
     void *p;
-    p = (void *)RtlDoSvc(190, (uintptr_t)size, 0, 0);
+    p = (void *)RtlDoSvc(190, (unsigned int)size, 0, 0);
     return p;
 }
 
@@ -13,13 +13,13 @@ void *RtlReallocateMemory(
     void *p,
     size_t size)
 {
-    RtlDoSvc(192, (uintptr_t)size, (uintptr_t)&p, 0);
+    RtlDoSvc(192, (unsigned int)size, (unsigned int)&p, 0);
     return p;
 }
 
 void RtlFreeMemory(
     void *p)
 {
-    RtlDoSvc(191, (uintptr_t)p, 0, 0);
+    RtlDoSvc(191, (unsigned int)p, 0, 0);
     return;
 }
