+++ dosrtl/debug.c
@@ -4,6 +4,6 @@
 void RtlDebugPrint(
     const char *str)
 {
-    RtlDoSvc(100, (uintptr_t)str, 0, 0);
+    RtlDoSvc(100, (unsigned int)str, 0, 0);
     return;
 }
